+++ src/main/java/codes/rudolph/ribac/DbHelper.java
@@ -1,0 +1,48 @@
+package codes.rudolph.ribac;
+
+import com.google.inject.Inject;
+import com.google.inject.name.Named;
+import io.reactivex.Scheduler;
+import io.reactivex.Single;
+import org.jooq.Configuration;
+import org.jooq.DSLContext;
+import org.jooq.impl.DSL;
+
+import java.util.function.Function;
+
+public class DbHelper {
+
+    private final Configuration jooqConfig;
+
+    private final Scheduler workerPool;
+
+
+
+    @Inject
+    public DbHelper(
+        Configuration jooqConfig,
+        @Named("workerPoolScheduler") Scheduler workerPool
+    ) {
+        this.jooqConfig = jooqConfig;
+        this.workerPool = workerPool;
+    }
+
+
+
+    /**
+     * Executes the given action on a worker pool thread inside a transaction and returns the result as a Single
+     *
+     * @param dbGetter The action to execute inside of the transaction. Must return a result.
+     * @param <R> The type of the result
+     * @return a single containing whatever the dbGetter returns
+     */
+    public <R> Single<R> execute(Function<DSLContext, R> dbGetter) {
+        return Single
+            .fromCallable(
+                () -> DSL.using(this.jooqConfig).transactionResult(
+                    transactionConfig -> dbGetter.apply(DSL.using(transactionConfig))
+                )
+            )
+            .subscribeOn(this.workerPool);
+    }
+}